// apps/web/src/lib/api/generated/main/model/notification-target-response-dto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { JsonValue } from './json-value';
import type { NotificationTargetResponseDtoOwnerType } from './notification-target-response-dto-owner-type';
import type { NotificationTargetResponseDtoProvider } from './notification-target-response-dto-provider';
import type { NotificationTargetResponseDtoTargetType } from './notification-target-response-dto-target-type';

export interface NotificationTargetResponseDto {
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  id: number;
  ownerType: NotificationTargetResponseDtoOwnerType;
  ownerId: string;
  provider: NotificationTargetResponseDtoProvider;
  targetType: NotificationTargetResponseDtoTargetType;
  externalId: string;
  /** @nullable */
  displayName: string | null;
  /** @nullable */
  guildName: string | null;
  metadata: (JsonValue | null) | null;
  active: boolean;
  canSend: boolean;
  /**
     * @nullable
     * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
     */
  lastSyncedAt: string | null;
  /**
     * @nullable
     * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
     */
  lastDeliveryAt: string | null;
  /** @nullable */
  lastDeliveryError: string | null;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  createdAt: string;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  updatedAt: string;
}